Move MediaFile-related files to new folder and sub-namespace
