+++ AudioTagger.Library/OutputLine.cs
@@ -1,3 +1,5 @@
+using AudioTagger.Library.MediaFiles;
+
 namespace AudioTagger;
 
 public class OutputLine
