+++ AudioTagger.Library/UpdatableFields.cs
@@ -1,5 +1,6 @@
 using System.Text.RegularExpressions;
 using System.Globalization;
+using AudioTagger.Library.MediaFiles;
 
 namespace AudioTagger;
